refactor: move root view to core
