+++ src/core/view/components/index.scss
