+++ src/core/view/components/index.tsx
