+++ src/core/view/components/toolbar/index.scss
